feat(document/controller): add document expiration cron job

+++ models/document.go
@@ -21,6 +21,6 @@ type Document struct {
 	ID        string `db:"id"`
 	Content   string `db:"content"`
 	Extension string `db:"extension"`
-	CreatedAt int    `db:"created_at"`
-	UpdatedAt int    `db:"updated_at"`
+	CreatedAt int64  `db:"created_at"`
+	UpdatedAt int64  `db:"updated_at"`
 }
