+++ structs/response.go
@@ -22,8 +22,8 @@ type Payload struct {
 	ID          *string `json:"id,omitempty"`           // The document ID.
 	Content     *string `json:"content,omitempty"`      // The document content.
 	Extension   *string `json:"extension,omitempty"`    // The extension of the document.
-	CreatedAt   *int    `json:"created_at,omitempty"`   // The Unix timestamp of when the document was inserted.
-	UpdatedAt   *int    `json:"updated_at,omitempty"`   // The Unix timestmap of when the document was last modified.
+	CreatedAt   *int64  `json:"created_at,omitempty"`   // The Unix timestamp of when the document was inserted.
+	UpdatedAt   *int64  `json:"updated_at,omitempty"`   // The Unix timestmap of when the document was last modified.
 	Exists      *bool   `json:"exists,omitempty"`       // Whether the document does or does not exist.
 }
